face-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0
tensorflow==2.13.0
keras==2.13.1
deepface==0.0.79
opencv-python-headless==4.8.0.74
tf-keras==2.16.0
